requirements.txt
# Required Python packages for ETL pipeline
psycopg2==2.9.2
pandas>=1.3.0,<2.0
numpy==1.21.1

# For loading environment variables from .env file
python-dotenv==0.19.0